+++ R/utils.R
@@ -10,3 +10,15 @@ modify_list <- function(old, new) {
   for (nm in names(new)) old[[nm]] <- new[[nm]]
   old
 }
+
+
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+#' Version number of 'yyjson' C library
+#' 
+#' @export
+#' @examples
+#' yyjson_version()
+#~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+yyjson_version <- function() {
+  .Call(yyjson_version_)
+}